pyproject.toml
[tool.poetry]
name = "garminworkouts"
version = "0.1.0"
description = "Create and schedule Garmin workouts"
authors = ["Alex Franz <alexfranz90@googlemail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.2"
PyYAML = "^6.0.1"
pandas = "^2.2.2"
garth = "^0.4.46"
typer = "^0.12.3"

[tool.poetry.group.dev.dependencies]
pytest_httpserver = "^1.0.10"
pre-commit = "^3.7.1"
pytest-mock = "^3.14.0"

[tool.poetry.scripts]
garminworkouts = "garminworkouts.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
